refactor: reorganiza pacotes de CategoriaCreateDTO e CategoriaCreateMapperImpl
